+++ sktree/tree/manifold/_morf_splitter.pxd
@@ -84,9 +84,9 @@ cdef class PatchSplitter(BaseObliqueSplitter):
     ) noexcept nogil
 
 
-cdef class UserKernelSplitter(PatchSplitter):
-    """A class to hold user-specified kernels."""
-    cdef vector[DTYPE_t[:, ::1]] kernel_dictionary  # A list of C-contiguous 2D kernels
+# cdef class UserKernelSplitter(PatchSplitter):
+#     """A class to hold user-specified kernels."""
+#     cdef vector[DTYPE_t[:, ::1]] kernel_dictionary  # A list of C-contiguous 2D kernels
 
 
 cdef class GaussianKernelSplitter(PatchSplitter):
