+++ sktree/tree/unsupervised/_unsup_splitter.pyx
@@ -174,7 +174,9 @@ cdef class BestUnsupervisedSplitter(UnsupervisedSplitter):
         self,
         double impurity,
         SplitRecord* split,
-        SIZE_t* n_constant_features
+        SIZE_t* n_constant_features,
+        double lower_bound,
+        double upper_bound
     ) except -1 nogil:
         """Find the best_split split on node samples[start:end].
